+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Feature
 
 - support to different types of routers
+- add fiber support
 
 ## 0.6.1 - 17-11-2022